debugging on hardware

+++ Apps/FlightSoftware/MotorControl/include/utils.h
@@ -6,7 +6,7 @@
 typedef struct { uint16_t  period;		// Input: Period of output impulse in a number of sampling cycles (Q0)
 				 uint8_t  cycle;      	// Output: 0x00 or 0xFF indicating if it is time to move to next commutation
 				                        //          cycle of motor
-		  	  	 uint16_t  counter;   	// Variable: Impulse generator counter (Q0)
+		  	  	 uint32_t  counter;   	// Variable: Impulse generator counter (Q0)
 			   } IMPULSE_TIMER;
